+++ docs/changes.md
@@ -10,7 +10,7 @@
 <br><br>
 * [ ] **Disable Evolution/Trade Restrictions**
   * [ ] Allow non-Kanto Pokémon to evolve by Level-Up
-  * [ ] Allow non-Kanto Pokémon to evolve with Trade Items
+  * [x] Allow non-Kanto Pokémon to evolve with Trade Items
   * [x] Allow non-Kanto Pokémon to evolve with Evolution Stones
   * [ ] Allow non-Kanto Pokémon to be traded
   * [ ] Allow Eggs to be traded